pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "reporting_service"
version = "0.1.0"
description = "Reporting Pipeline for a Simulated Broker Example."
readme = "README.md"
requires-python = ">=3.8"
license = { file = "LICENSE" }
authors = [
  { name = "Your Name", email = "your.email@example.com" }
]
dependencies = [
  "pyspark==3.5.5", 
  "delta-spark==3.1.0",
  "dbt-core==1.8",
  "dbt-duckdb==1.8",
  "duckdb==0.10.*",
  "rich",
  "click",
  "pytest",
  "jinja2",
  "plotly", 
  "pandas"
]
classifiers = [
  "Programming Language :: Python :: 3",
  "Operating System :: OS Independent"
]


[tool.setuptools]
# Tells setuptools where your packages actually live
package-dir = {"" = "src"}


[tool.setuptools.packages.find]
where = ["src", ""]
include = ["pipeline", "sb_pipeline"]
exclude = ["logs*", "tests*", "docs*"]
